requirements.txt
# requirements.txt
flask==2.0.1
pika==1.2.0
requests==2.26.0
pandas==1.3.5
stable-baselines3==1.3.0
gym==0.21.0
numpy==1.21.4
transformers==4.12.5
sqlalchemy==1.4.27
redis==4.0.2
python-dotenv==0.19.2